// apps/desktop/src/generated/RuntimeEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RuntimeEvent =
  | { type: "output"; runtimeId: string; data: Array<number> }
  | { type: "error"; runtimeId: string; message: string }
  | { type: "exit"; runtimeId: string; code: number };